add permissions to views

+++ app_answer/views.py
@@ -4,6 +4,8 @@ from rest_framework.response import Response
 from rest_framework.views import APIView
 from .models import Answer
 from .serializers import AnswerSerializer
+from permissions import IsOwnerOrReadOnly
+from rest_framework.permissions import IsAuthenticated
 
 
 class ListAnswerView(APIView):
@@ -14,6 +16,8 @@ class ListAnswerView(APIView):
 
 
 class CreateAnswerView(APIView):
+    permission_classes = [IsAuthenticated]
+
     def post(self, request):
         srz_data = AnswerSerializer(data=request.data)
         if srz_data.is_valid():
@@ -24,8 +28,11 @@ class CreateAnswerView(APIView):
 
 
 class UpdateAnswerView(APIView):
+    permission_classes = [IsOwnerOrReadOnly]
+
     def patch(self, request, pk):
         answer = get_object_or_404(Answer, pk=pk)
+        self.check_object_permissions(request, answer)
         srz_data = AnswerSerializer(instance=answer, data=request.data, partial=True)
         if srz_data.is_valid():
             srz_data.save()
@@ -35,7 +42,10 @@ class UpdateAnswerView(APIView):
 
 
 class DeleteAnswerView(APIView):
+    permission_classes = [IsOwnerOrReadOnly]
+
     def delete(self, request, pk):
         answer = get_object_or_404(Answer, pk=pk)
+        self.check_object_permissions(request, answer)
         answer.delete()
         return Response(data={'message': 'answer deleted successfully!'}, status=status.HTTP_200_OK)
