+++ app_bookmark/views.py
@@ -4,6 +4,8 @@ from rest_framework.response import Response
 from rest_framework.views import APIView
 from .models import Bookmark
 from .serializers import BookmarkSerializer
+from permissions import IsOwnerOrReadOnly
+from rest_framework.permissions import IsAuthenticated
 
 
 class ListBookmarkView(APIView):
@@ -12,7 +14,10 @@ class ListBookmarkView(APIView):
         srz_data = BookmarkSerializer(instance=bookmarks, many=True)
         return Response(srz_data.data, status=status.HTTP_200_OK)
 
+
 class CreateBookmarkView(APIView):
+    permission_classes = [IsAuthenticated]
+
     def post(self, request):
         srz_data = BookmarkSerializer(data=request.data)
         if srz_data.is_valid():
@@ -23,8 +28,11 @@ class CreateBookmarkView(APIView):
 
 
 class UpdateBookmarkView(APIView):
+    permission_classes = [IsOwnerOrReadOnly]
+
     def patch(self, request, pk):
         bookmark = get_object_or_404(Bookmark, pk=pk)
+        self.check_object_permissions(request, bookmark)
         srz_data = BookmarkSerializer(instance=bookmark, data=request.data, partial=True)
         if srz_data.is_valid():
             srz_data.save()
@@ -34,7 +42,10 @@ class UpdateBookmarkView(APIView):
 
 
 class DeleteBookmarkView(APIView):
+    permission_classes = [IsOwnerOrReadOnly]
+
     def delete(self, request, pk):
         bookmark = get_object_or_404(Bookmark, pk=pk)
+        self.check_object_permissions(request, bookmark)
         bookmark.delete()
         return Response(data={'message': 'bookmark deleted successfully!'}, status=status.HTTP_200_OK)
