+++ app_comment/views.py
@@ -4,6 +4,8 @@ from rest_framework.response import Response
 from rest_framework.views import APIView
 from .models import Comment
 from .serializers import CommentSerializer
+from permissions import IsOwnerOrReadOnly
+from rest_framework.permissions import IsAuthenticated
 
 
 class ListCommentView(APIView):
@@ -14,6 +16,8 @@ class ListCommentView(APIView):
 
 
 class CreateCommentView(APIView):
+    permission_classes = [IsAuthenticated]
+
     def post(self, request):
         srz_data = CommentSerializer(data=request.data)
         if srz_data.is_valid():
@@ -24,8 +28,11 @@ class CreateCommentView(APIView):
 
 
 class UpdateCommentView(APIView):
+    permission_classes = [IsOwnerOrReadOnly]
+
     def patch(self, request, pk):
         comment = get_object_or_404(Comment, pk=pk)
+        self.check_object_permissions(request, comment)
         srz_data = CommentSerializer(instance=comment, data=request.data, partial=True)
         if srz_data.is_valid():
             srz_data.save()
@@ -35,7 +42,10 @@ class UpdateCommentView(APIView):
 
 
 class DeleteCommentView(APIView):
+    permission_classes = [IsOwnerOrReadOnly]
+
     def delete(self, request, pk):
         comment = get_object_or_404(Comment, pk=pk)
+        self.check_object_permissions(request, comment)
         comment.delete()
         return Response(data={'message': 'comment deleted successfully!'}, status=status.HTTP_200_OK)
