+++ app_question/views.py
@@ -4,6 +4,8 @@ from rest_framework.response import Response
 from rest_framework.views import APIView
 from .models import Question
 from .serializers import QuestionSerializer
+from permissions import IsOwnerOrReadOnly
+from rest_framework.permissions import IsAuthenticated
 
 
 class ListQuestionView(APIView):
@@ -14,6 +16,8 @@ class ListQuestionView(APIView):
 
 
 class CreateQuestionView(APIView):
+    permission_classes = [IsAuthenticated]
+
     def post(self, request):
         srz_data = QuestionSerializer(data=request.data)
         if srz_data.is_valid():
@@ -24,8 +28,11 @@ class CreateQuestionView(APIView):
 
 
 class UpdateQuestionView(APIView):
+    permission_classes = [IsOwnerOrReadOnly]
+
     def patch(self, request, pk):
         answer = get_object_or_404(Question, pk=pk)
+        self.check_object_permissions(request, answer)
         srz_data = QuestionSerializer(instance=answer, data=request.data, partial=True)
         if srz_data.is_valid():
             srz_data.save()
@@ -35,8 +42,11 @@ class UpdateQuestionView(APIView):
 
 
 class DeleteQuestionView(APIView):
+    permission_classes = [IsOwnerOrReadOnly]
+
     def delete(self, request, pk):
         answer = get_object_or_404(Question, pk=pk)
+        self.check_object_permissions(request, answer)
         answer.delete()
         return Response(data={'message': 'answer deleted successfully!'}, status=status.HTTP_200_OK)
