+++ app_tag/views.py
@@ -4,6 +4,8 @@ from django.shortcuts import get_object_or_404
 from rest_framework import status
 from rest_framework.views import APIView
 from rest_framework.response import Response
+from permissions import IsOwnerOrReadOnly
+from rest_framework.permissions import IsAuthenticated
 
 
 class ListTagView(APIView):
@@ -14,6 +16,8 @@ class ListTagView(APIView):
 
 
 class CreateTagView(APIView):
+    permission_classes = [IsAuthenticated]
+
     def post(self, request):
         srz_data = TagSerializer(data=request.data)
         if srz_data.is_valid():
@@ -24,8 +28,11 @@ class CreateTagView(APIView):
 
 
 class UpdateTagView(APIView):
+    permission_classes = [IsOwnerOrReadOnly]
+
     def patch(self, request, pk):
         tag = get_object_or_404(Tag, pk=pk)
+        self.check_object_permissions(request, tag)
         srz_data = TagSerializer(instance=tag, data=request.data, partial=True)
         if srz_data.is_valid():
             srz_data.save()
@@ -35,7 +42,10 @@ class UpdateTagView(APIView):
 
 
 class DeleteTagView(APIView):
+    permission_classes = [IsOwnerOrReadOnly]
+
     def delete(self, request, pk):
         tag = get_object_or_404(Tag, pk=pk)
+        self.check_object_permissions(request, tag)
         tag.delete()
         return Response(data={'message': f'tag {tag.name} deleted successfully!'}, status=status.HTTP_204_NO_CONTENT)
